(fix) Improve typing of store actions
