+++ packages/framework/esm-react-utils/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./ComponentContext";
 export * from "./ConfigurableLink";
-export * from "./createUseStore";
 export * from "./Extension";
 export * from "./ExtensionSlot";
 export * from "./getLifecycle";
