+++ packages/framework/esm-react-utils/src/useExtensionStore.ts
@@ -1,6 +1,6 @@
 /** @module @category Extension */
 import { ExtensionStore, getExtensionStore } from "@openmrs/esm-extensions";
-import { createUseStore } from "./createUseStore";
+import { createUseStore } from "./useStore";
 
 export const useExtensionStore = createUseStore<ExtensionStore>(
   getExtensionStore()
