+++ packages/framework/esm-react-utils/src/useStore.ts
@@ -2,9 +2,14 @@
 import { subscribeTo } from "@openmrs/esm-state";
 import { useEffect, useMemo, useState } from "react";
 import type { StoreApi } from "zustand";
-import type { Actions, BoundActions } from "./createUseStore";
 
-function bindActions<T>(store: StoreApi<T>, actions: Actions): BoundActions {
+export type ActionFunction<T> = (state: T, ...args: any[]) => Partial<T>;
+export type Actions<T> =
+  | ((store: StoreApi<T>) => Record<string, ActionFunction<T>>)
+  | Record<string, ActionFunction<T>>;
+export type BoundActions = { [key: string]: (...args: any[]) => void };
+
+function bindActions<T>(store: StoreApi<T>, actions: Actions<T>): BoundActions {
   if (typeof actions == "function") {
     actions = actions(store);
   }
@@ -35,17 +40,17 @@ function useStore<T, U>(store: StoreApi<T>, select: (state: T) => U): U;
 function useStore<T, U>(
   store: StoreApi<T>,
   select: undefined,
-  actions: Actions
+  actions: Actions<T>
 ): T & BoundActions;
 function useStore<T, U>(
   store: StoreApi<T>,
   select: (state: T) => U,
-  actions: Actions
+  actions: Actions<T>
 ): U & BoundActions;
 function useStore<T, U>(
   store: StoreApi<T>,
   select: (state: T) => U = defaultSelectFunction,
-  actions?: Actions
+  actions?: Actions<T>
 ) {
   const [state, setState] = useState<U>(() => select(store.getState()));
   useEffect(() => subscribeTo(store, select, setState), [store, select]);
@@ -58,11 +63,39 @@ function useStore<T, U>(
   return { ...state, ...boundActions };
 }
 
+/**
+ *
+ * @param store A zustand store
+ * @param actions
+ * @returns
+ */
 function useStoreWithActions<T>(
   store: StoreApi<T>,
-  actions: Actions
+  actions: Actions<T>
 ): T & BoundActions {
   return useStore(store, defaultSelectFunction, actions);
 }
 
-export { useStore, useStoreWithActions };
+/**
+ * Whenever possible, use `useStore(yourStore)` instead. This function is for creating a
+ * custom hook for a specific store.
+ */
+function createUseStore<T>(store: StoreApi<T>) {
+  function useStore(): T;
+  function useStore(actions: Actions<T>): T & BoundActions;
+  function useStore(actions?: Actions<T>): T & BoundActions;
+  function useStore(actions?: Actions<T>) {
+    const [state, set] = useState(store.getState());
+    useEffect(() => store.subscribe((state) => set(state)), []);
+    let boundActions: BoundActions = useMemo(
+      () => (actions ? bindActions(store, actions) : {}),
+      [actions]
+    );
+
+    return { ...state, ...boundActions };
+  }
+
+  return useStore;
+}
+
+export { createUseStore, useStore, useStoreWithActions };
